+++ lib/ephemeral_bloc.dart
@@ -1,5 +1,6 @@
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/ephemeral_bloc_change.dart';
+export 'src/ephemeral_bloc_consumer.dart';
+export 'src/ephemeral_bloc_listener.dart';
+export 'src/ephemeral_bloc_mixin.dart';
+export 'src/ephemeral_bloc_observer.dart';
+export 'src/ephemeral_bloc_streamable.dart';